# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Very Useful Tools to Remember",
    "description": "API consumed by server to register and manage tools"
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/",
  "paths": {
    "/tools": {
      "get": {
        "summary": "Gets all tools",
        "tags": [
          "VUTTR"
        ],
        "description": "Returns a list containing all tools.",
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "500": {
            "description": "500 Internal server error"
          }
        }
      },
      "post": {
        "summary": "Registers a new tool.",
        "tags": [
          "VUTTR"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "tool",
            "description": "The tool to create.",
            "schema": {
              "type": "object",
              "required": [
                "title",
                "description",
                "link",
                "tags"
              ],
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "link": {
                  "type": "string"
                },
                "tags": {
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "description": "Returns the recently registered tool with its _id.",
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "500": {
            "description": "500 Internal server error"
          }
        }
      }
    },
    "/tools/{_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "summary": "Updates a tool based on its _id.",
        "tags": [
          "VUTTR"
        ],
        "description": "Returns the recently updated tool.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tool",
            "in": "body",
            "schema": {
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "link": {
                  "type": "string"
                },
                "tags": {
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "500": {
            "description": "500 Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Deletes a tool based on its _id.",
        "tags": [
          "VUTTR"
        ],
        "description": "Returns an empty object.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "500": {
            "description": "500 Internal server error"
          }
        }
      },
      "get": {
        "summary": "Gets a tool based on its _id.",
        "tags": [
          "VUTTR"
        ],
        "description": "Returns the tool data.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "200 OK"
          },
          "500": {
            "description": "500 Internal server error"
          }
        }
      }
    }
  }
}